+++ mimic-bukkit/src/test/kotlin/impl/skillapi/SkillApiConverterTest.kt
@@ -19,16 +19,24 @@
 
 package ru.endlesscode.mimic.impl.skillapi
 
+import io.kotest.matchers.shouldBe
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.Arguments.arguments
 import org.junit.jupiter.params.provider.MethodSource
 import ru.endlesscode.mimic.level.ExpLevelConverter
 import java.util.stream.Stream
-import kotlin.test.BeforeTest
-import kotlin.test.assertEquals
 
 class SkillApiConverterTest : SkillApiTestBase() {
+
+    // SUT
+    private val converter: ExpLevelConverter = SkillApiConverter.getInstance(skillApi)
+
+    @ParameterizedTest
+    @MethodSource("expLevel")
+    fun testExpToLevel(exp: Double, level: Double) {
+        converter.expToLevel(exp) shouldBe level
+    }
 
     @Suppress("unused")
     companion object {
@@ -39,25 +47,5 @@ class SkillApiConverterTest : SkillApiTestBase() {
             arguments(100.0, 5.0),
             arguments(140.0, 5.8)
         )
-    }
-
-    // SUT
-    private lateinit var converter: ExpLevelConverter
-
-    @BeforeTest
-    override fun setUp() {
-        super.setUp()
-
-        converter = SkillApiConverter.getInstance(skillApi)
-    }
-
-    @ParameterizedTest
-    @MethodSource("expLevel")
-    fun testExpToLevel(exp: Double, level: Double) {
-        // When
-        val actualLevel = converter.expToLevel(exp)
-
-        // Then
-        assertEquals(level, actualLevel)
     }
 }